quality-selector styles & cleanup
