+++ src/assets/styles/main.scss
@@ -10,7 +10,7 @@ $icon-font-path: '../fonts' !default;
 @import 'icons';
 @import 'ads-label';
 @import 'videojs-ima';
-@import '~videojs-hls-quality-selector/dist/videojs-hls-quality-selector.css';
+// @import '~videojs-hls-quality-selector/dist/videojs-hls-quality-selector.css';
 
 // control bar control size
 .video-js .vjs-control:before {
@@ -264,17 +264,22 @@ $icon-font-path: '../fonts' !default;
   }
 
   .vjs-subs-caps-button {
-    .vjs-menu {
-      width: auto;
-      left: -5em;
+    .vjs-texttrack-settings {
+      display: none;
     }
+    .vjs-captions-menu-item .vjs-menu-item-text .vjs-icon-placeholder {
+      vertical-align: top;
+      display: inline-block;
+      margin-bottom: -0.3em;
+    }
+  }
+
+  .vjs-menu {
     .vjs-menu-content {
+      min-width: 10em;
       width: auto;
       overflow: visible;
       padding: 0.2em 0;
-    }
-    .vjs-texttrack-settings {
-      display: none;
     }
     .vjs-menu-item {
       text-align: left;
@@ -285,11 +290,6 @@ $icon-font-path: '../fonts' !default;
       line-height: 2em;
       &.vjs-selected {
         font-weight: bold;
-      }
-      &.vjs-captions-menu-item .vjs-menu-item-text .vjs-icon-placeholder {
-        vertical-align: top;
-        display: inline-block;
-        margin-bottom: -0.3em;
       }
     }
   }
