+++ src/components/qualitySelector/quality-selector.scss
@@ -1,0 +1,7 @@
+.video-js {
+  .vjs-http-source-selector {
+    .vjs-button {
+      font-size: 17px;
+    }
+  }
+}
