+++ src/plugins/index.js
@@ -4,9 +4,7 @@ import './ima';
 import dashPlugin from './dash';
 import interactive from './interactive-plugin';
 // #endif
-import 'videojs-per-source-behaviors';
-import 'videojs-contrib-quality-levels';
-import 'videojs-http-source-selector';
+
 import autoplayOnScroll from './autoplay-on-scroll';
 import floatingPlayer from './floating-player';
 import colors from './colors';
